make() and pop() debugging, print_list() attempt
~ doubly_linked_list.c
	+ added main for debugging
	~ fixed pop not setting the popped neighbour's prev or next to NULL
	+ shallow debugging for make() and pop()

	+ added a print_list() function for debugging but it doesnt work for some reason lmao
	+ now importing inttypes.h for print_list() even tho it doesnt work asiocjaocja

~ doubly_linked_list.h
	~ size is now int instead of int64_t

LEGEND:
+ : added
~ : modified

+++ doubly-linked-list/doubly_linked_list.h
@@ -11,7 +11,7 @@
 typedef struct list {
     node* left;
     node* right;
-    int64_t size;
+    int size;
     bool is_reversed;
 } list;
